Added documentation.
