+++ tests/test_password.py
@@ -1,11 +1,14 @@
 from argparse import Namespace
-from unittest import TestCase
 from awsauthhelper.password import generate
+from tests import AuthHelperTestCase
 
 __author__ = 'drews'
 
 
-class TestPassword(TestCase):
+class TestPassword(AuthHelperTestCase):
+    """
+    Make sure that our password generation has enough entropy.
+    """
     SYMBOLS = list("!@#$%^&*()_+-=[]{}|'")
     LOWERCASE = list("abcdefghijklmnopqrstuvwxyz")
     UPPERCASE = list("ABCDEFGHIJKLMNOPQRSTUVWXYZ")
@@ -29,46 +32,46 @@ class TestPassword(TestCase):
 
         # Test override
         new_password = generate(self.MockPasswordPolicy(
-            minimum_password_length=10
+                minimum_password_length=10
         ))
         self.assertGreaterEqual(len(new_password), 10)
         self.assertIsInstance(new_password, basestring)
 
     def test_generate_symbols(self):
         new_password = generate(self.MockPasswordPolicy(
-            require_symbols=True
+                require_symbols=True
         ))
         self.assertAnyIn(self.SYMBOLS, new_password)
         self.assertIsInstance(new_password, basestring)
 
     def test_generate_lowercase(self):
         new_password = generate(self.MockPasswordPolicy(
-            require_lowercase_characters=True
+                require_lowercase_characters=True
         ))
         self.assertAnyIn(self.LOWERCASE, new_password)
         self.assertIsInstance(new_password, basestring)
 
     def test_generate_uppercase(self):
         new_password = generate(self.MockPasswordPolicy(
-            require_uppercase_characters=True
+                require_uppercase_characters=True
         ))
         self.assertAnyIn(self.UPPERCASE, new_password)
         self.assertIsInstance(new_password, basestring)
 
     def test_generate_numbers(self):
         new_password = generate(self.MockPasswordPolicy(
-            require_numbers=True
+                require_numbers=True
         ))
         self.assertAnyIn(self.NUMBERS, new_password)
         self.assertIsInstance(new_password, basestring)
 
     def test_generate_secure(self):
         new_password = generate(self.MockPasswordPolicy(
-            minimum_password_length=12,
-            require_symbols=True,
-            require_lowercase_characters=True,
-            require_uppercase_characters=True,
-            require_numbers=True
+                minimum_password_length=12,
+                require_symbols=True,
+                require_lowercase_characters=True,
+                require_uppercase_characters=True,
+                require_numbers=True
         ))
         self.assertGreaterEqual(len(new_password), 12)
         self.assertAnyIn(self.SYMBOLS, new_password)
@@ -76,10 +79,3 @@ class TestPassword(TestCase):
         self.assertAnyIn(self.UPPERCASE, new_password)
         self.assertAnyIn(self.NUMBERS, new_password)
         self.assertIsInstance(new_password, basestring)
-
-    def assertAnyIn(self, needles, haystack):
-        if not any(x in haystack for x in needles):
-            raise AssertionError('None of \'{needles}\' in \'{haystack}\''.format(
-                needles=",".join(needles),
-                haystack=haystack
-            ))
